WF-27 generator
